changed `unsigned` to `size_t`
`size_t` plays nicely overall than `unsigned` in the algorithm used.

+++ Cycle-1/1_fcfs.c
@@ -46,7 +46,7 @@ size_t schedule(Process *process, size_t n)
 
     size_t z = 0;    /* The current chart element. */
     size_t i = 0;    /* The current process from table. */
-    unsigned t = 0;  /* The current time. */
+    size_t t = 0;  /* The current time. */
     bool busy = false;
     Process p;
